Add files via upload
Project to Create a Database Schema and Table Relationships for a Logistic Company's Data.

The logistics company provides services in both the international and domestic sectors. The logistics management takes into consideration every facility that has an impact on
cost. It plays an important role in making the product confirm to customer requirements. Also, it involves efficient integration of suppliers, manufacturers, Import & export and
other activities at many levels; from the strategic level through the tactical to the operational level. Customers can send different types of shipping contents. Payment is to be done at the same time the product is delivered to the client. Delivery boy and center head can update the status of the shipment. Create a database schema and table relationships that can be used in any technology.